% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeCommunities.R
\name{align2refGenome}
\alias{align2refGenome}
\title{Aligner of FastQ files with reference genome}
\usage{
align2refGenome(
  filt_files,
  genome,
  aligner = "Rbowtie",
  pair_end = F,
  parameters,
  nclusters = 4
)
}
\arguments{
\item{filt_files}{A filtered FastQ file or a list with filtered FastQ files, derived from microbiomeTrimAndQuality()}

\item{genome}{A Fasta file or a BSgenome object, as a reference genome}

\item{aligner}{A String indicating the preferred aligner, between "Rbowtie" or "Hisat2"}

\item{pair_end}{A logical value if the data are single or pair-end (TRUE is for pair-end)}

\item{parameters}{Additional parameters as the aligner support}

\item{nclusters}{The number of processes to be created that each points to 1 core}
}
\value{
The BAM and BAI files of the alignment, in the same dir as input files
}
\description{
Aligner of FastQ files with reference genome
}
