% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{analyzeLeaderlessRNA}
\alias{analyzeLeaderlessRNA}
\title{Analyze Leaderless mRNAs}
\usage{
analyzeLeaderlessRNA(TSSdataframe, TSS_5UTR, GeneAnnotation)
}
\arguments{
\item{TSSdataframe}{A Dataframe with all the TSS information}

\item{TSS_5UTR}{A Dataframe derived from getRNA5UTRsequencies function}

\item{GeneAnnotation}{A 2 column Dataframe containing basic annotation for each gene}
}
\value{
A Dataframe containing leaderless mRNAs with its annotation
}
\description{
Analyze Leaderless mRNAs
}
