% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeCommunities.R
\name{analyzeTSSmicrobes}
\alias{analyzeTSSmicrobes}
\title{Analyze BED files}
\usage{
analyzeTSSmicrobes(bamfiles, cutoff = 10)
}
\arguments{
\item{bamfiles}{An string path or a list of paths, representing different Replicates, of BAM files}

\item{cutoff}{A numeric value of clustering distance}
}
\value{
A Dataframe with TSS information
}
\description{
Analyze BED files
}
