% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{annotateRNA}
\alias{annotateRNA}
\title{Annotate RNAs to coding (mRNA) and non-coding (other RNA)}
\usage{
annotateRNA(TSSdataframe, nc_RNA_path)
}
\arguments{
\item{TSSdataframe}{A Dataframe containing start, strand and RRS information for each TSS}

\item{nc_RNA_path}{A string path containing a tab-separated list of non-coding RNAs}
}
\value{
A list of RNA annotation to mRNA or non-coding
}
\description{
Annotate RNAs to coding (mRNA) and non-coding (other RNA)
}
