% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operons.R
\name{assignOperons}
\alias{assignOperons}
\title{Assign Operons to TSS}
\usage{
assignOperons(TSSdataframe, operons)
}
\arguments{
\item{TSSdataframe}{A Dataframe containing start, strand, gene, position and type information for each TSS}

\item{operons}{A Dataframe containing operon information, including operon names, genes, positions etc.}
}
\value{
A final Dataframe with the assigned operon name to each TSS and the number of genes that contains
}
\description{
Assign Operons to TSS
}
