% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeCommunities.R
\name{assignTaxonmies}
\alias{assignTaxonmies}
\title{Assign taxonomies using 16S rRNA gene classifier}
\usage{
assignTaxonmies(filt_fastq, rdp_train_set, pool, plot = T, multithread = T)
}
\arguments{
\item{filt_fastq}{A String path to trimed and filtered FastQ file, or a list of these files}

\item{rdp_train_set}{A RDP Database train set of 16S rRNA genes}

\item{pool}{A logical value to merge into a pool all the data of the input FastQ files}

\item{plot}{A logical value to show (TRUE) or not show (FALSE) plots}

\item{multithread}{A logical value to use (TRUE) or not (FALSE) multithreading}
}
\value{
A Dataframe with the classified reads
}
\description{
Assign taxonomies using 16S rRNA gene classifier
}
