% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAM2firstbaseBED.R
\name{bam2bedTSS}
\alias{bam2bedTSS}
\title{BAM to BED conversion, keeping the first base of each read from the 5' end}
\usage{
bam2bedTSS(bamFile)
}
\arguments{
\item{bamFile}{A BamFile object, from the package Rsamtools}
}
\value{
A Dataframe containing the first base of each read (candidate TSS), based on its strand
}
\description{
BAM to BED conversion, keeping the first base of each read from the 5' end
}
