% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAM2FIrstBaseBAM.R
\name{BAM2firstbaseBAM}
\alias{BAM2firstbaseBAM}
\title{Convert BAM fileto first base BAM, keeping the first base from the 5' end of each read}
\usage{
BAM2firstbaseBAM(bamFile, genomeFai)
}
\arguments{
\item{bamFile}{A BAM file string path or a BAM Dataframe}

\item{genomeFai}{A FAI file of the reference genome}
}
\value{
A first base BAM file as a file
}
\description{
Convert BAM fileto first base BAM, keeping the first base from the 5' end of each read
}
