% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{betaBinomial}
\alias{betaBinomial}
\title{Calculate p-values using Beta Binomial model, based on the RSS scores of 2 samples}
\usage{
betaBinomial(
  RRSenriched,
  RRScontrol,
  TotalEnriched = 1e+06,
  TotalControl = 1e+06,
  ncores = round(detectCores()/3, 0),
  plot = T
)
}
\arguments{
\item{RRSenriched}{A list or Dataframe column, containing RSS scores from the enriched Cappable-seq sample}

\item{RRScontrol}{A list or Dataframe column, containing RSS scores from the control Cappable-seq sample}

\item{TotalEnriched}{A numeric value, indicating the sum of RSS scores in the enriched sample, typically 1e6}

\item{TotalControl}{A numeric value, indicating the sum of RSS scores in the enriched sample, typically 1e6}

\item{ncores}{The number of cores to parallel the calculation}

\item{plot}{A logical value to show (TRUE) or not show (FALSE) plots}
}
\value{
A list containing all the p-values of TSS
}
\description{
Calculate p-values using Beta Binomial model, based on the RSS scores of 2 samples
}
