% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSScluster.R
\name{clusterTSS}
\alias{clusterTSS}
\title{Cluster neighboring TSS based on cutoff value}
\usage{
clusterTSS(TSSenriched, cutoff = 5)
}
\arguments{
\item{TSSenriched}{A Dataframe of the unclustered TSS}

\item{cutoff}{A numeric value of the cluster distance}
}
\value{
A clustered Dataframe
}
\description{
Cluster neighboring TSS based on cutoff value
}
