% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeCommunities.R
\name{createNJtree}
\alias{createNJtree}
\title{Create Neighbor Joining tree and plot microbiome composition}
\usage{
createNJtree(taxonomies, seqtable, plot = T)
}
\arguments{
\item{taxonomies}{Taxonomies object derived from assignTaxonomies function (taxa identifier)}

\item{seqtable}{Sequencies object with removed Chimeras derived from assignTaxonomies function (noChim identifier)}

\item{plot}{A logical value to show (TRUE) or not show (FALSE) plots}
}
\value{
Phyloseq object, phylogenetic tree and microbiome composition plots
}
\description{
Create Neighbor Joining tree and plot microbiome composition
}
