% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{dinucleotideTSS}
\alias{dinucleotideTSS}
\title{Calculate and plot DiNucleotide Frequencies (-1 to +1)}
\usage{
dinucleotideTSS(start, strand, RRSenriched, genome, plots = T)
}
\arguments{
\item{start}{A list or a Dataframe column, containing the start genomic coordinates of TSS}

\item{strand}{A list or a Dataframe column, containing the strand genomic information of each TSS}

\item{RRSenriched}{A list or a Dataframe column, containing RSS scores from the enriched Cappable-seq sample}

\item{genome}{The reference genome in format of readDNAStringSet or BSgenome object}

\item{plots}{A logical value to show (TRUE) or not show (FALSE) plots}
}
\value{
A Dataframe containing the DiNucleotide, withcategorization to Purine/Pyrimidine
}
\description{
Calculate and plot DiNucleotide Frequencies (-1 to +1)
}
