% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{fisherExactTest}
\alias{fisherExactTest}
\title{Adjust P-values with Fisher Exact Test}
\usage{
fisherExactTest(pValue)
}
\arguments{
\item{pValue}{A Dataframe with p-values from sample's replicates}
}
\value{
A Dataframe with Adjusted p-values
}
\description{
Adjust P-values with Fisher Exact Test
}
