% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{getPromoterSequencies}
\alias{getPromoterSequencies}
\title{Get Promoter sequences using a reference genome}
\usage{
getPromoterSequencies(start, strand, genome, left = 40, right = 5, plot = T)
}
\arguments{
\item{start}{A list or a Dataframe column, containing the start genomic coordinates of TSS}

\item{strand}{A list or a Dataframe column, containing the strand genomic information of each TSS}

\item{genome}{The reference genome in format of readDNAStringSet or BSgenome object}

\item{left}{A numeric integer and positive value for the left range from TSS}

\item{right}{A numeric integer and positive value for the right range from TSS}

\item{plot}{A logical value to show (TRUE) or not show (FALSE) plots}
}
\value{
A Dataframe with the selected promoter sequencies
}
\description{
Get Promoter sequences using a reference genome
}
