% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{getRepresentativeTSS}
\alias{getRepresentativeTSS}
\title{Get Upstream Representative TSS}
\usage{
getRepresentativeTSS(TSSdataframe)
}
\arguments{
\item{TSSdataframe}{An annotated TSS Dataframe produced from annotateTSS and betaBinomial functions}
}
\value{
A Dataframe containing all Upstream Representative TSS
}
\description{
Get Upstream Representative TSS
}
