% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{getRNA5UTRsequencies}
\alias{getRNA5UTRsequencies}
\title{Get Nucleotide Sequences from TSS to Start Codon of mRNAs}
\usage{
getRNA5UTRsequencies(TSSdataframe, genome, GeneList, plot = T)
}
\arguments{
\item{TSSdataframe}{A Dataframe containing start, strand and orientation, position and type information for each TSS}

\item{genome}{The reference genome in format of readDNAStringSet or BSgenome object}

\item{GeneList}{A Dataframe containing gene names, left & right coordinates and strand information}

\item{plot}{A logical value to show (TRUE) or not show (FALSE) plots}
}
\value{
A Dataframe containing 5' UTR sequencies, Gene start position and TSS position
}
\description{
Get Nucleotide Sequences from TSS to Start Codon of mRNAs
}
