% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{getShineDalgarnoSequencies}
\alias{getShineDalgarnoSequencies}
\title{Get and plot Shine-Dalgarno genomic region}
\usage{
getShineDalgarnoSequencies(TSS_5UTR, plot = T)
}
\arguments{
\item{TSS_5UTR}{A Dataframe derived from getRNA5UTRsequencies function}

\item{plot}{A logical value to show (TRUE) or not show (FALSE) plots}
}
\value{
A Dataframe containing sequencies from 15 bases upstream of TSS to TSS, for each one
}
\description{
Get and plot Shine-Dalgarno genomic region
}
