% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeCommunities.R
\name{microbiomeTrimAndQuality}
\alias{microbiomeTrimAndQuality}
\title{Trim and filter FastQ files}
\usage{
microbiomeTrimAndQuality(
  mi_path,
  adapters,
  minLength,
  trucnLength,
  minQuality,
  trancQuality,
  ncluster = 4
)
}
\arguments{
\item{mi_path}{A string path of only the FastQ files to be analyzed}

\item{adapters}{Adapter sequence from the 3' end to be trimmed}

\item{minLength}{Minimum length of trimmed reads}

\item{trucnLength}{Minimum length of reads to be truncated}

\item{minQuality}{Minimum quality of trimmed reads}

\item{trancQuality}{Minimum quality of reads to be truncated}

\item{ncluster}{An integer value, indicating the number of clusters (process)}
}
\value{
Trimed and filtered FastQ files and 2 quality plots
}
\description{
Trim and filter FastQ files
}
