% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operons.R
\name{plotOperonPreference}
\alias{plotOperonPreference}
\title{Plot TSS preference inside Operons and distances upstream of Operon start site}
\usage{
plotOperonPreference(TSSdataframe, operons)
}
\arguments{
\item{TSSdataframe}{A Dataframe containing start, strand, gene, position and type information for each TSS}

\item{operons}{A Dataframe containing operon information, including operon names, genes, positions etc.}
}
\value{
2 Plots of TSS preference inside Operons and distances upstream of Operon start site
}
\description{
Plot TSS preference inside Operons and distances upstream of Operon start site
}
