% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{plotTSSperCategory}
\alias{plotTSSperCategory}
\title{Logo Plots of all TSS categories}
\usage{
plotTSSperCategory(TSSdataframe, genome)
}
\arguments{
\item{TSSdataframe}{A Dataframe containing start, strand and orientation, position and type information for each TSS}

\item{genome}{The reference genome in format of readDNAStringSet or BSgenome object}
}
\value{
A logo plot of all TSS categories
}
\description{
Logo Plots of all TSS categories
}
