% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{plotTSSpreference}
\alias{plotTSSpreference}
\title{Plot internal TSS preference information}
\usage{
plotTSSpreference(TSSdataframe, GeneList)
}
\arguments{
\item{TSSdataframe}{A Dataframe containing start, strand and orientation, position and type information for each TSS}

\item{GeneList}{A Dataframe containing gene names, left & right coordinates and strand information}
}
\value{
2 plots. One of internal TSS preference inside genes and other of codon-based position of TSS related with protein coding genes
}
\description{
Plot internal TSS preference information
}
