% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKnowledge.R
\name{smallRNAanalysis}
\alias{smallRNAanalysis}
\title{Title}
\usage{
smallRNAanalysis(TSSdataframe, sRNAfile)
}
\arguments{
\item{TSSdataframe}{An annotated TSS Dataframe produced from annotateRNA function}

\item{sRNAfile}{A string path containing a tab-separated list of non-coding RNAs}
}
\value{
The dataframe with the TSS, which correspond to small RNAs
}
\description{
Title
}
