% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrepocess.R
\name{trimAlignAndQuality}
\alias{trimAlignAndQuality}
\title{Sequencing Data Preprocess, including filter, trim and alignment with the reference genome}
\usage{
trimAlignAndQuality(
  fastq,
  adapter,
  minLength,
  refGenome,
  aligner,
  complexity = 0.4,
  isPaired,
  ncores
)
}
\arguments{
\item{fastq}{A FastQ file or a list with FastQ files to analyze}

\item{adapter}{Adapter raw sequence}

\item{minLength}{A number indicating the minimum length of reads}

\item{refGenome}{A Fasta file or a BSgenome object, as a reference genome}

\item{aligner}{A String indicating the preferred aligner, between "Rbowtie" or "Hisat2"}

\item{complexity}{A number indicating the minimum entropy value, according to human complexity}

\item{isPaired}{A logical value if the data are single or pair-end}

\item{ncores}{The number of processes to be created that each points to 1 core}
}
\value{
The BAM, BAI files of the alignment, as well as a PDF file as a report
}
\description{
Sequencing Data Preprocess, including filter, trim and alignment with the reference genome
}
